# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_create_submission_api_v1_submissions_post": {
        "properties": {
          "contact_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Email"
          },
          "context": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context"
          },
          "files": {
            "default": [],
            "items": {
              "contentMediaType": "application/octet-stream",
              "type": "string"
            },
            "title": "Files",
            "type": "array"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          }
        },
        "title": "Body_create_submission_api_v1_submissions_post",
        "type": "object"
      },
      "DecisionFields": {
        "description": "Opportunity payload a moderator supplies when approving a submission.",
        "properties": {
          "category": {
            "enum": [
              "scholarship",
              "internship",
              "competition"
            ],
            "title": "Category",
            "type": "string"
          },
          "deadline": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deadline"
          },
          "description": {
            "minLength": 1,
            "title": "Description",
            "type": "string"
          },
          "eligibility": {
            "minLength": 1,
            "title": "Eligibility",
            "type": "string"
          },
          "issuer_name": {
            "anyOf": [
              {
                "maxLength": 160,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Issuer Name"
          },
          "region": {
            "anyOf": [
              {
                "maxLength": 160,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region"
          },
          "slug": {
            "anyOf": [
              {
                "maxLength": 160,
                "pattern": "^[a-z0-9]+(-[a-z0-9]+)*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slug"
          },
          "title": {
            "maxLength": 240,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title",
          "category",
          "description",
          "eligibility"
        ],
        "title": "DecisionFields",
        "type": "object"
      },
      "DecisionRequest": {
        "properties": {
          "decision": {
            "enum": [
              "approved",
              "needs_more_evidence",
              "rejected",
              "expire"
            ],
            "title": "Decision",
            "type": "string"
          },
          "fields": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DecisionFields"
              },
              {
                "type": "null"
              }
            ]
          },
          "reason": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "required": [
          "decision"
        ],
        "title": "DecisionRequest",
        "type": "object"
      },
      "DecisionResult": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "opportunity_slug": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Opportunity Slug"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "submission_id": {
            "format": "uuid",
            "title": "Submission Id",
            "type": "string"
          },
          "submission_state": {
            "title": "Submission State",
            "type": "string"
          }
        },
        "required": [
          "id",
          "status",
          "submission_id",
          "submission_state"
        ],
        "title": "DecisionResult",
        "type": "object"
      },
      "DecisionView": {
        "properties": {
          "actor_id": {
            "format": "uuid",
            "title": "Actor Id",
            "type": "string"
          },
          "decided_at": {
            "format": "date-time",
            "title": "Decided At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "status",
          "reason",
          "actor_id",
          "decided_at"
        ],
        "title": "DecisionView",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "QueueItem": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "has_contact_email": {
            "title": "Has Contact Email",
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "open_reports_count": {
            "title": "Open Reports Count",
            "type": "integer"
          },
          "ref": {
            "title": "Ref",
            "type": "string"
          },
          "state": {
            "title": "State",
            "type": "string"
          },
          "submitted_url_host": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submitted Url Host"
          },
          "uploads_count": {
            "title": "Uploads Count",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "ref",
          "state",
          "created_at",
          "submitted_url_host",
          "uploads_count",
          "has_contact_email",
          "open_reports_count"
        ],
        "title": "QueueItem",
        "type": "object"
      },
      "QueueResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/QueueItem"
            },
            "title": "Items",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total"
        ],
        "title": "QueueResponse",
        "type": "object"
      },
...
      "ReportView": {
        "properties": {
          "category": {
            "title": "Category",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "category",
          "description",
          "status",
          "created_at"
        ],
        "title": "ReportView",
        "type": "object"
      },
      "ScreeningRunMeta": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "model_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Version"
          },
          "provider_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Version"
          },
          "schema_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Schema Version"
          },
          "state": {
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "id",
          "state",
          "provider_version",
          "model_version",
          "schema_version",
          "created_at"
        ],
        "title": "ScreeningRunMeta",
        "type": "object"
      },
      "SubmissionCreated": {
        "properties": {
          "receipt_token": {
            "title": "Receipt Token",
            "type": "string"
          },
          "ref": {
            "title": "Ref",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "status_url": {
            "title": "Status Url",
            "type": "string"
          }
        },
        "required": [
          "ref",
          "receipt_token",
          "status",
          "status_url"
        ],
        "title": "SubmissionCreated",
        "type": "object"
      },
      "SubmissionDetail": {
        "description": "Everything a moderator sees; receipt_token_hash is never included.",
        "properties": {
          "client_net_hash": {
            "title": "Client Net Hash",
            "type": "string"
          },
          "contact_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Email"
          },
          "context": {
            "title": "Context",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "decisions": {
            "items": {
              "$ref": "#/components/schemas/DecisionView"
            },
            "title": "Decisions",
            "type": "array"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "linked_opportunity_slug": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Linked Opportunity Slug"
          },
          "purge_after": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Purge After"
          },
          "ref": {
            "title": "Ref",
            "type": "string"
          },
          "reports": {
            "items": {
              "$ref": "#/components/schemas/ReportView"
            },
            "title": "Reports",
            "type": "array"
          },
          "screening_run": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ScreeningRunMeta"
              },
              {
                "type": "null"
              }
            ]
          },
          "state": {
            "title": "State",
            "type": "string"
          },
          "submitted_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submitted Url"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "uploads": {
            "items": {
              "$ref": "#/components/schemas/UploadMeta"
            },
            "title": "Uploads",
            "type": "array"
          }
        },
        "required": [
          "id",
          "ref",
          "state",
          "submitted_url",
          "context",
          "contact_email",
          "client_net_hash",
          "purge_after",
          "created_at",
          "updated_at",
          "uploads",
          "screening_run",
          "decisions",
          "linked_opportunity_slug",
          "reports"
        ],
        "title": "SubmissionDetail",
        "type": "object"
      },
      "SubmissionStatus": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "decision": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision"
          },
          "needs_more_evidence": {
            "default": false,
            "title": "Needs More Evidence",
            "type": "boolean"
          },
          "ref": {
            "title": "Ref",
            "type": "string"
          },
          "state": {
            "title": "State",
            "type": "string"
          },
          "status_label": {
            "title": "Status Label",
            "type": "string"
          }
        },
        "required": [
          "ref",
          "state",
          "status_label",
          "created_at"
        ],
        "title": "SubmissionStatus",
        "type": "object"
      },
      "UploadMeta": {
        "properties": {
          "detected_mime": {
            "title": "Detected Mime",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "page_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Count"
          },
          "size_bytes": {
            "title": "Size Bytes",
            "type": "integer"
          },
          "storage_key": {
            "title": "Storage Key",
            "type": "string"
          }
        },
        "required": [
          "id",
          "storage_key",
          "detected_mime",
          "size_bytes",
          "page_count"
        ],
        "title": "UploadMeta",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "APIKeyCookie": {
        "in": "cookie",
        "name": "jp_auth",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "JejakPeluang API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/moderation/submissions": {
      "get": {
        "operationId": "list_submissions_api_v1_moderation_submissions_get",
        "parameters": [
          {
            "in": "query",
            "name": "state",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueueResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "List Submissions",
        "tags": [
          "moderation"
        ]
      }
    },
    "/api/v1/moderation/submissions/{submission_id}": {
      "get": {
        "operationId": "get_submission_detail_api_v1_moderation_submissions__submission_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "submission_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Submission Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmissionDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Get Submission Detail",
        "tags": [
          "moderation"
        ]
      }
    },
    "/api/v1/moderation/submissions/{submission_id}/decision": {
      "post": {
        "operationId": "post_decision_api_v1_moderation_submissions__submission_id__decision_post",
        "parameters": [
          {
            "in": "path",
            "name": "submission_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Submission Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Post Decision",
        "tags": [
          "moderation"
        ]
      }
    },
    "/api/v1/moderation/submissions/{submission_id}/uploads/{upload_id}": {
      "get": {
        "description": "Stream a stored upload for moderator review.\n\nThe upload row is looked up by id and must belong to the submission;\nupload_file_path resolves storage_key under UPLOAD_DIR and refuses any\npath outside it, so no user input ever reaches the filesystem.",
        "operationId": "download_upload_api_v1_moderation_submissions__submission_id__uploads__upload_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "submission_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Submission Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "upload_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Upload Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Download Upload",
        "tags": [
          "moderation"
        ]
      }
    },
...
    "/api/v1/submissions": {
      "post": {
        "operationId": "create_submission_api_v1_submissions_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_submission_api_v1_submissions_post"
              }
            }
          }
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmissionCreated"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Submission",
        "tags": [
          "submissions"
        ]
      }
    },
    "/api/v1/submissions/{ref}": {
      "get": {
        "operationId": "get_submission_status_api_v1_submissions__ref__get",
        "parameters": [
          {
            "in": "path",
            "name": "ref",
            "required": true,
            "schema": {
              "title": "Ref",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-receipt-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Receipt-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmissionStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Submission Status",
        "tags": [
          "submissions"
        ]
      }
    },
...
  }
}